edit-tab-again * Move addNewStation functionality into BikePanelSearch * Remove position and onSelected from BikePanelSearch props
