+++ src/containers/Admin/EditTab/BikeTile/BikeSearch/BikePanelSearch.tsx
@@ -5,8 +5,8 @@ import {
     BikePanelSearchStationFragment,
     useBikePanelSearchQuery,
 } from '../../../../../../graphql-generated/mobility-v2'
-import { Coordinates } from '../../../../../types'
 import { isNotNullOrUndefined } from '../../../../../utils/typeguards'
+import { useSettings } from '../../../../../settings/SettingsProvider'
 import classes from './BikePanelSearch.module.scss'
 
 const MAX_SEARCH_RANGE = 100_000
@@ -25,47 +25,54 @@ function mapFeaturesToItems(
     }))
 }
 
-interface BikePanelSearchProps {
-    onSelected: (stationId: string) => void
-    position: Coordinates
-}
+const BikePanelSearch: React.FC = () => {
+    const [settings, setSettings] = useSettings()
 
-const BikePanelSearch: React.FC<BikePanelSearchProps> = ({
-    onSelected,
-    position,
-}) => {
     const { data } = useBikePanelSearchQuery({
         variables: {
-            lat: position?.latitude,
-            lon: position?.longitude,
+            lat: settings.coordinates.latitude,
+            lon: settings.coordinates.longitude,
             range: MAX_SEARCH_RANGE,
         },
         fetchPolicy: 'cache-and-network',
     })
 
-    const getItems = (query: string): Item[] => {
-        const inputValue = query.trim().toLowerCase()
-        const inputLength = inputValue.length
-        if (!inputLength || !data?.stations) return []
+    const getItems = useCallback(
+        (query: string): Item[] => {
+            const inputValue = query.trim().toLowerCase()
+            const inputLength = inputValue.length
+            if (!inputLength || !data?.stations) return []
 
-        return mapFeaturesToItems(
-            data.stations
-                .filter(isNotNullOrUndefined)
-                .filter((station) =>
-                    getTranslation(station?.name)
-                        ?.toLowerCase()
-                        .match(new RegExp(inputValue)),
-                ),
-        )
-    }
+            return mapFeaturesToItems(
+                data.stations
+                    .filter(isNotNullOrUndefined)
+                    .filter((station) =>
+                        getTranslation(station?.name)
+                            ?.toLowerCase()
+                            .match(new RegExp(inputValue)),
+                    ),
+            )
+        },
+        [data?.stations],
+    )
+
+    const addNewStation = useCallback(
+        (stationId: string) => {
+            if (settings.newStations.includes(stationId)) return
+            setSettings({
+                newStations: [...settings.newStations, stationId],
+            })
+        },
+        [settings.newStations, setSettings],
+    )
 
     const handleOnChange = useCallback(
         (item: Item | null): void => {
             if (item) {
-                onSelected(item.value)
+                addNewStation(item.value)
             }
         },
-        [onSelected],
+        [addNewStation],
     )
 
     return (
